Refactored code, updated documentation
I reformated the code with an object oriented view. Attached the uml
class diagramm too.

This way the code is easier to understand (even if it became
significantly longer +30%)

The Plan is to support constructors like argumentsAsInParent and
argumentsSimilatToFields. The oop approach will allow me to do this
quite easily